Add files via upload
The codes for the Fig. 5, 6, and 7.
1. RT&IAE Calculation
We employed two parameters from control theory, the response time (RT) and the integral of absolute error (IAE), to characterize the dynamics of the QS-regulated circular-interaction microbial consortium. Codes in this file were used for the Fig. 5 and Supplementary Fig. 2-3.

2. Chage R for different typology
We investigated the potential effect from the one-way changeable QS-regulated microbial interactions imposed on strains A, B, and E, respectively. 
Codes in this file were used for the Fig. 6 and Supplementary Fig. 5.

3. SAforD
As three strains in the consortium will compete on space and nutrition, which will depend on the relative concentrations of strain cells, controlling of operation conditions like dilution rate (D) will be very important for maintaining the stability of the consortium. Therefore, we conducted an analysis for D (being 0.01, 0.05, 0.076, 0.1, 0.134, or 0.3) on the cell growth of strains E, A, and B. 
Codes in this file were used for the Fig. 7a-7c.

4. SAforDandR
We conducted a sensitivity analysis for two operating parameters, the promoted cell growth rate and D, which are both set in the range from 0.01 to 0.3. 
Codes in this file were used for the Fig. 7d-7e.

5. Initial ratio
In addition, we also tested the sensitivity of the consortium stability on the initial inoculation ratio of three strains.
Codes in this file were used for the Fig. 7f-7g and Supplementary Fig. 6.